move module tests files
